+++ .env.production.example
@@ -95,6 +95,31 @@ STREAM_SEGMENT_CACHE_TTL_SECONDS=3600
 STREAM_SEGMENT_CACHE_MAX_BYTES=67108864
 STREAM_REQUEST_TIMEOUT_SECONDS=60
 
+# Optional monitoring/logging overlay settings.
+PROMETHEUS_IMAGE=prom/prometheus:latest
+PROMETHEUS_HTTP_PORT=9090
+ALERTMANAGER_IMAGE=prom/alertmanager:latest
+ALERTMANAGER_HTTP_PORT=9093
+GRAFANA_IMAGE=grafana/grafana:latest
+GRAFANA_HTTP_PORT=3000
+GRAFANA_ADMIN_USER=admin
+GRAFANA_ADMIN_PASSWORD=change-me-grafana-admin-password
+GRAFANA_AUTH_ANONYMOUS_ENABLED=false
+LOKI_IMAGE=grafana/loki:2.9.8
+LOKI_HTTP_PORT=3100
+PROMTAIL_IMAGE=grafana/promtail:2.9.8
+PROMTAIL_HTTP_PORT=9080
+
+# Paste these into private Alertmanager config snippets or use them when
+# rendering a host-local monitoring/alertmanager/alertmanager.yml.
+ALERTMANAGER_SLACK_WEBHOOK_URL=
+ALERTMANAGER_SMTP_SMARTHOST=
+ALERTMANAGER_SMTP_FROM=
+ALERTMANAGER_SMTP_USERNAME=
+ALERTMANAGER_SMTP_PASSWORD=
+ALERTMANAGER_EMAIL_TO=
+ALERTMANAGER_PAGERDUTY_ROUTING_KEY=
+
 # Local/custom EVM payment settings. Leave blank for standard network config.
 PROD_EVM_RPC_URL=
 PROD_EVM_PAYMENT_TOKEN_ADDRESS=
