+++ Makefile
@@ -1,8 +1,15 @@
-.PHONY: help install-react test test-rust test-rust-workspace test-rust-stream test-rust-admin test-rust-db test-antd clippy-rust clippy-rust-workspace clippy-rust-stream clippy-rust-admin clippy-antd fmt-rust compose-config backup-production restore-production test-react build-react smoke-local smoke-local-restart smoke-local-large-original audit-rust audit-react audit-trivy audit ci
+.PHONY: help install-react test test-rust test-rust-workspace test-rust-stream test-rust-admin test-rust-db test-antd clippy-rust clippy-rust-workspace clippy-rust-stream clippy-rust-admin clippy-antd fmt-rust compose-config up-local up-local-full up-prod down-local down-prod logs logs-prod logs-monitoring backup-production restore-production test-react build-react smoke-local smoke-local-restart smoke-local-large-original audit-rust audit-react audit-trivy audit ci
 
 NPM ?= npm
 CARGO ?= cargo
 DOCKER_COMPOSE ?= docker compose
+LOCAL_ENV ?= .env.local
+PROD_ENV ?= .env.production
+LOCAL_COMPOSE_FILES = -f docker-compose.yml -f docker-compose.local.yml
+LOCAL_MONITORING_COMPOSE_FILES = $(LOCAL_COMPOSE_FILES) -f docker-compose.monitoring.yml
+PROD_COMPOSE_FILES = -f docker-compose.yml -f docker-compose.prod.yml
+CORE_LOG_SERVICES = rust_admin rust_stream antd nginx react_frontend db
+MONITORING_LOG_SERVICES = prometheus alertmanager grafana
 
 help:
 	@echo "Available targets:"
@@ -19,6 +26,14 @@ help:
 	@echo "  make clippy-antd     Run antd service clippy checks"
 	@echo "  make fmt-rust        Check Rust formatting"
 	@echo "  make compose-config  Validate Compose render paths"
+	@echo "  make up-local        Start local devnet stack in the background"
+	@echo "  make up-local-full   Start local devnet stack with monitoring"
+	@echo "  make up-prod         Start production/default-network stack"
+	@echo "  make down-local      Stop local devnet stack without deleting volumes"
+	@echo "  make down-prod       Stop production stack without deleting volumes"
+	@echo "  make logs            Follow local core service logs"
+	@echo "  make logs-prod       Follow production core service logs"
+	@echo "  make logs-monitoring Follow local monitoring service logs"
 	@echo "  make backup-production Create a timestamped production DB/catalog backup"
 	@echo "  make restore-production ARGS='--backup-dir backups/autvid-... --yes' Restore a production backup"
 	@echo "  make install-react   Install React frontend dependencies"
@@ -73,6 +88,31 @@ compose-config:
 	$(DOCKER_COMPOSE) --env-file .env.local.example -f docker-compose.yml -f docker-compose.local.yml config >/tmp/autvid-compose-local.yml
 	$(DOCKER_COMPOSE) --env-file .env.local-public.example -f docker-compose.yml -f docker-compose.local.yml -f docker-compose.local-public.yml config >/tmp/autvid-compose-local-public.yml
 	$(DOCKER_COMPOSE) --env-file .env.production.example -f docker-compose.yml -f docker-compose.prod.yml config >/tmp/autvid-compose-prod.yml
+	$(DOCKER_COMPOSE) --env-file .env.production.example -f docker-compose.yml -f docker-compose.prod.yml -f docker-compose.monitoring.yml -f docker-compose.logging.yml config >/tmp/autvid-compose-prod-observability.yml
+
+up-local:
+	$(DOCKER_COMPOSE) --env-file $(LOCAL_ENV) $(LOCAL_COMPOSE_FILES) up --build -d
+
+up-local-full:
+	$(DOCKER_COMPOSE) --env-file $(LOCAL_ENV) $(LOCAL_MONITORING_COMPOSE_FILES) up --build -d
+
+up-prod:
+	$(DOCKER_COMPOSE) --env-file $(PROD_ENV) $(PROD_COMPOSE_FILES) up --build -d
+
+down-local:
+	$(DOCKER_COMPOSE) --env-file $(LOCAL_ENV) $(LOCAL_COMPOSE_FILES) down
+
+down-prod:
+	$(DOCKER_COMPOSE) --env-file $(PROD_ENV) $(PROD_COMPOSE_FILES) down
+
+logs:
+	$(DOCKER_COMPOSE) --env-file $(LOCAL_ENV) $(LOCAL_COMPOSE_FILES) logs -f $(CORE_LOG_SERVICES)
+
+logs-prod:
+	$(DOCKER_COMPOSE) --env-file $(PROD_ENV) $(PROD_COMPOSE_FILES) logs -f $(CORE_LOG_SERVICES)
+
+logs-monitoring:
+	$(DOCKER_COMPOSE) --env-file $(LOCAL_ENV) $(LOCAL_MONITORING_COMPOSE_FILES) logs -f $(MONITORING_LOG_SERVICES)
 
 backup-production:
 	scripts/backup-production.sh
