+++ monitoring/promtail/config.yml
@@ -1,0 +1,29 @@
+server:
+  http_listen_port: 9080
+  grpc_listen_port: 0
+
+positions:
+  filename: /run/promtail/positions.yaml
+
+clients:
+  - url: http://loki:3100/loki/api/v1/push
+
+scrape_configs:
+  - job_name: docker
+    docker_sd_configs:
+      - host: unix:///var/run/docker.sock
+        refresh_interval: 10s
+    relabel_configs:
+      - source_labels:
+          - __meta_docker_container_name
+        regex: "/(.*)"
+        target_label: container
+      - source_labels:
+          - __meta_docker_container_label_com_docker_compose_service
+        target_label: service
+      - source_labels:
+          - __meta_docker_container_label_com_docker_compose_project
+        target_label: compose_project
+      - source_labels:
+          - __meta_docker_container_label_com_docker_compose_config_hash
+        target_label: compose_config_hash
