+++ nginx/conf.d/default.conf
@@ -1,5 +1,10 @@
 limit_req_zone $binary_remote_addr zone=autvid_login:10m rate=5r/m;
 
+map $http_x_forwarded_proto $autvid_hsts {
+    default "";
+    ~*^https$ "max-age=63072000; includeSubDomains; preload";
+}
+
 server {
     listen 8080;
     server_name _;
@@ -12,6 +17,9 @@ server {
     add_header X-Content-Type-Options nosniff always;
     add_header X-Frame-Options DENY always;
     add_header Referrer-Policy no-referrer always;
+    add_header Strict-Transport-Security $autvid_hsts always;
+    add_header Content-Security-Policy "default-src 'self'; script-src 'self'; style-src 'self' 'unsafe-inline'; img-src 'self' data: blob:; media-src 'self' blob:; connect-src 'self'; worker-src 'self' blob:; object-src 'none'; base-uri 'self'; frame-ancestors 'none'" always;
+    add_header Permissions-Policy "camera=(), microphone=(), geolocation=()" always;
 
     location / {
         set                $frontend_backend http://react_app:8080;
