+++ react_frontend/src/hooks/useAuth.ts
@@ -1,15 +1,47 @@
 import { useCallback, useEffect, useRef, useState } from "react";
 
-import { getCurrentUser, logoutAdmin, refreshAdmin } from "../api/client";
+import { getCurrentUser, logoutAdmin, refreshAdmin, subscribeAuthRefresh } from "../api/client";
 import { AUTH_STORAGE_KEY } from "../constants";
 import type { AuthState } from "../types";
 
 export default function useAuth(onInvalid?: () => void) {
   const skipNextRefresh = useRef(false);
+  const onInvalidRef = useRef(onInvalid);
+  const authRef = useRef<AuthState | null>(null);
   const [auth, setAuth] = useState<AuthState | null>(() => {
     const token = window.localStorage.getItem(AUTH_STORAGE_KEY);
-    return token ? { access_token: token, username: "" } : null;
+    const initialAuth = token ? { access_token: token, username: "" } : null;
+    authRef.current = initialAuth;
+    return initialAuth;
   });
+
+  useEffect(() => {
+    onInvalidRef.current = onInvalid;
+  }, [onInvalid]);
+
+  useEffect(() => {
+    authRef.current = auth;
+  }, [auth]);
+
+  useEffect(() => subscribeAuthRefresh((nextAuth) => {
+    if (nextAuth) {
+      const refreshedAuth = {
+        ...nextAuth,
+        username: nextAuth.username || authRef.current?.username || "",
+      };
+      skipNextRefresh.current = false;
+      authRef.current = refreshedAuth;
+      setAuth(refreshedAuth);
+      return;
+    }
+
+    const hadAuth = !!authRef.current?.access_token;
+    skipNextRefresh.current = true;
+    authRef.current = null;
+    window.localStorage.removeItem(AUTH_STORAGE_KEY);
+    setAuth(null);
+    if (hadAuth) onInvalidRef.current?.();
+  }), []);
 
   useEffect(() => {
     if (!auth?.access_token && skipNextRefresh.current) {
@@ -44,6 +76,7 @@ export default function useAuth(onInvalid?: () => void) {
         }
         window.localStorage.removeItem(AUTH_STORAGE_KEY);
         if (active) {
+          authRef.current = null;
           setAuth(null);
           if (auth?.access_token) onInvalid?.();
         }
@@ -58,11 +91,13 @@ export default function useAuth(onInvalid?: () => void) {
   }, [auth?.access_token, onInvalid]);
 
   const login = useCallback((nextAuth: AuthState) => {
+    authRef.current = nextAuth;
     setAuth(nextAuth);
   }, []);
 
   const logout = useCallback(() => {
     skipNextRefresh.current = true;
+    authRef.current = null;
     logoutAdmin().catch(() => {});
     setAuth(null);
   }, []);
